Ignore events other than play
All events in gPodder are :
- play
- download
- delete
- new
- flattr
(see https://gpoddernet.readthedocs.io/en/latest/api/reference/events.html)

`download` and `delete` are ignored by AntennaPod and cause issues with some frontends (see https://github.com/AntennaPod/AntennaPod/issues/6957#issuecomment-2050513363)

For `new`, the official documentaion says "This state needs to be interpreted by receiving clients and does not delete any information on the webservice." However, gpodder-sync will erase position and total if client put some value in it (-1 most of the time). It does not respect the spec and should be ignored as well.

`flattr` is undocumented and seems unused.

By elimination, only play is really helpfull.
We should not remove the action column because it can have NEW, DOWNLOAD and DELETE action already with -1 data and could cause damage if removed.